Can calculate median of sorted and unsorted arrays

+++ src/VectorUtils.hpp
@@ -8,8 +8,8 @@ namespace Project5 {
 namespace Vector {
 
 double calculateMean(const std::vector<double> &values) {
-  return std::accumulate(values.begin(), values.end(), 0) /
-         (double)values.size();
+  return (double)std::accumulate(values.begin(), values.end(), (double)0) /
+         values.size();
 }
 
 double calculateMedianSorted(const std::vector<double> &values) {
@@ -22,7 +22,8 @@ double calculateMedianSorted(const std::vector<double> &values) {
   } else {
     std::array<double, 2> middle = {values[values.size() / 2 - 1],
                                     values[values.size() / 2]};
-    return std::accumulate(middle.begin(), middle.end(), 0) / middle.size();
+    return (double)std::accumulate(middle.begin(), middle.end(), (double)0) /
+           middle.size();
   }
 }
 } // namespace Vector
